stm32: Add casts when printing small integers.
All these arguments are of type `mp_{u,}int_t`, but the actual value is
always a small integer.  Cast it so that it can format with the `%d/%u`
formatter.

Before, the compiler plugin produced an error in the PYBD_SF6 build, which
is a nanboxing build with 64-bit ints.

+++ ports/stm32/extint.c
@@ -760,7 +760,7 @@ static mp_obj_t extint_make_new(const mp_obj_type_t *type, size_t n_args, size_t
 
 static void extint_obj_print(const mp_print_t *print, mp_obj_t self_in, mp_print_kind_t kind) {
     extint_obj_t *self = MP_OBJ_TO_PTR(self_in);
-    mp_printf(print, "<ExtInt line=%u>", self->line);
+    mp_printf(print, "<ExtInt line=%u>", (int)self->line);
 }
 
 static const mp_rom_map_elem_t extint_locals_dict_table[] = {
